@@ -20,64 +20,62 @@
   },
   "main": "dist/index.js",
   "scripts": {
-    "prepare": "nps prepare",
+    "prepare": "nr prepare",
     "prepublishOnly": "re-pack guard",
-    "help": "nps --scripts"
+    "help": "nr --scripts"
   },
   "dependencies": {
-    "@babel/preset-react": "^7.13.13",
+    "@babel/plugin-transform-runtime": "^7.14.5",
+    "@babel/preset-react": "^7.14.5",
     "@darkobits/chex": "^1.0.1",
-    "@darkobits/log": "^2.0.0-beta.15",
-    "@darkobits/ts": "^0.3.3",
+    "@darkobits/log": "^2.0.0-beta.16",
+    "@darkobits/ts": "^0.7.6",
     "@hot-loader/react-dom": "^17.0.1",
     "@svgr/webpack": "^5.5.0",
-    "@types/bytes": "^3.1.0",
-    "@types/copy-webpack-plugin": "^6.4.0",
-    "@types/friendly-errors-webpack-plugin": "^0.1.2",
-    "@types/mini-css-extract-plugin": "^1.2.2",
+    "@types/bytes": "^3.1.1",
+    "@types/copy-webpack-plugin": "^8.0.1",
+    "@types/mini-css-extract-plugin": "^2.0.1",
     "@types/ms": "^0.7.31",
-    "@types/node": "^14.14.22",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/webpack-dev-server": "^3.11.0",
+    "@types/node": "^16.4.0",
+    "@types/react": "^17.0.14",
+    "@types/react-dom": "^17.0.9",
+    "@types/webpack-dev-server": "^3.11.5",
     "babel-loader": "^8.2.2",
     "bytes": "^3.1.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^6.4.1",
-    "core-js": "^3.12.1",
-    "css-loader": "^5.0.1",
-    "dotenv": "^9.0.2",
+    "copy-webpack-plugin": "^9.0.1",
+    "core-js": "^3.15.2",
+    "css-loader": "^6.2.0",
+    "dotenv": "^10.0.0",
     "eslint-loader": "^4.0.2",
-    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-hooks": "^4.0.8",
-    "favicons-webpack-plugin": "^4.2.0",
     "file-loader": "^6.1.0",
     "find-up": "^5.0.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.7",
-    "friendly-errors-webpack-plugin": "^1.7.0",
+    "fork-ts-checker-webpack-plugin": "^6.2.12",
     "fs-extra": "^10.0.0",
-    "html-webpack-plugin": "^4.5.1",
+    "html-webpack-plugin": "^5.3.2",
     "is-ci": "^3.0.0",
     "linaria": "^2.0.2",
-    "mini-css-extract-plugin": "^1.6.0",
+    "mini-css-extract-plugin": "^2.1.0",
     "ms": "^2.1.3",
-    "postcss": "^8.2.4",
+    "postcss": "^8.3.6",
     "raw-loader": "^4.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hot-loader": "^4.12.21",
     "regenerator-runtime": "^0.13.7",
     "resolve-pkg": "^2.0.0",
-    "typescript-styled-plugin": "^0.17.0",
+    "typescript-styled-plugin": "^0.18.0",
     "url-loader": "^4.1.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.11",
+    "webpack": "^5.45.1",
+    "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   },
   "devDependencies": {
-    "@darkobits/re-pack": "^0.1.11",
-    "@types/fs-extra": "^9.0.11",
+    "@darkobits/nr": "^0.7.0",
+    "@darkobits/re-pack": "^0.1.12",
+    "@types/fs-extra": "^9.0.12",
     "@types/is-ci": "^3.0.0"
   }
 }
