@@ -1,1 +1,5 @@
-module.exports = require('@darkobits/ts').jest();
+import { jest } from '@darkobits/ts';
+
+export default jest();
+
+// module.exports = require('@darkobits/ts').jest();
